Moved SerializedXZY from Timeline to Persistence, created one for snapshot
